Gems/AWSCore/cdkv1/requirements.txt
aws-cdk.core>=1.91.0
aws-cdk.aws_iam>=1.91.0
aws-cdk.aws_resourcegroups>=1.91.0

aws-cdk.aws_s3>=1.91.0
aws-cdk.aws_s3_deployment>=1.91.0
aws-cdk.aws_lambda>=1.91.0
aws_cdk.aws_dynamodb>=1.91.0

boto3>=1.17.31

